+++ tests/unit/runtime/events/fund_transfer_event_1.json
@@ -10,7 +10,7 @@
     "customer_id": "CUS_01FWWSK432VY3X1T8A4VNYRTGR",
     "account_id": "ACC_01FWWSNA9DA3N3EQ2JHPQ4WTNR",
     "transaction_id": "gesNQKWBXUDKJKgJqKjrsg",
-    "amount": 1234.56,
+    "transfer_amount": 1234.56,
     "currency": "SGD",
     "memo": "to some random account",
     "transaction_date": "2022-02-27",
